+++ frontend/src/components/Review/index.ts
@@ -1,4 +1,5 @@
 export { default as ReviewItem } from './ReviewItem/ReviewItem';
+export { default as ReviewList } from './ReviewList/ReviewList';
 export { default as ReviewTagItem } from './ReviewTagItem/ReviewTagItem';
 export { default as ReviewTagList } from './ReviewTagList/ReviewTagList';
 export { default as ReviewRegisterForm } from './ReviewRegisterForm/ReviewRegisterForm';
