+++ frontend/src/hooks/product/index.ts
@@ -3,3 +3,4 @@ export { default as useCategoryProducts } from './useCategoryProducts';
 export { default as useProductReview } from './useProductReview';
 export { default as useProductDetail } from './useProductDetail';
 export { default as useInfiniteProducts } from './useInfiniteProducts';
+export { default as useInfiniteProductReviews } from './useInfiniteProductReviews';
